--- a/build/Dockerfile
+++ b/build/Dockerfile
@@ -2,7 +2,7 @@
 # See LICENSE.txt for license information.
 
 # Build the mattermost cloud
-ARG DOCKER_BUILD_IMAGE=golang:1.22
+ARG DOCKER_BUILD_IMAGE=golang:1.24
 ARG DOCKER_BASE_IMAGE=alpine:3.19
 
 FROM --platform=${TARGETPLATFORM} ${DOCKER_BUILD_IMAGE} AS build