--- a/scripts/push-image-pr.sh
+++ b/scripts/push-image-pr.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-set -e
-set -u
-
-make build-image